requirements.txt
amqp
billiard
celery
Django==4.1.3
django-filter
django-viewflow
pytz
# redis==2.10.6
kombu
vine

